refactor: $lib/agent
