+++ src/lib/agent/contentProposal.ts
@@ -39,45 +39,3 @@ export async function parseContentProposal(llmResponse: string): Promise<Content
 		reason: reason || undefined
 	};
 }
-
-/**
- * Generiert User-Prompt für Phase 2 (Structure Generation)
- * 
- * @param content - Der genehmigte Content aus Phase 1
- * @param existingColumns - Bestehende Spalten (falls vorhanden)
- * @returns User-Prompt String für LLM
- */
-export function generateStructurePrompt(content: string, existingColumns: string[] = []): string {
-	const hasExistingColumns = existingColumns.length > 0;
-
-	let prompt = `Erstelle eine Kanban-Board-Struktur basierend auf folgendem Inhalt:\n\n${content}\n\n`;
-
-	if (hasExistingColumns) {
-		prompt += `WICHTIG: Das Board hat bereits folgende Spalten:\n${existingColumns.map((col) => `- ${col}`).join('\n')}\n\n`;
-		prompt += `Füge NUR neue Karten zu diesen bestehenden Spalten hinzu. Erstelle KEINE neuen Spalten.\n\n`;
-	}
-
-	prompt += `Antworte NUR mit einem validen JSON-Objekt im folgenden Format (ohne Code-Blöcke):\n\n`;
-	prompt += JSON.stringify(
-		{
-			title: 'Board Titel',
-			description: 'Optional: Board Beschreibung',
-			columns: [
-				{
-					name: 'Spaltenname',
-					cards: [
-						{
-							heading: 'Karten-Titel',
-							content: 'Optional: Karten-Beschreibung',
-							labels: ['label1', 'label2']
-						}
-					]
-				}
-			]
-		},
-		null,
-		2
-	);
-
-	return prompt;
-}
